Clunky first implementation of LET in FUNC/FUNCTION
R3-Alpha introduced the idea of a "FUNCT", which would notice when a
SET-WORD! was used in a function body that was not one of the arguments
or locals in the spec.  It would then automatically "gather" these and
add them as locals in the spec.

This was much more convenient than having to maintain by hand a list
of the locals in the spec, which might get out of date as the body was
edited.  The convenience was such that this concept took the name
FUNCTION away from it former use (as a slight variant of FUNC which
let you separate the locals into their own parameter block).

However, the idea of auto-gathered locals based on SET-WORD! runs into
several problems in Rebol--as SET-WORD!s are used for many purposes
besides local variables.  Function frames became inflated with stray
variables that were really just keys in objects being used, and there
were many mechanical consequences.  It also led to FUNC having a
meaning that was distinct from an abbreviation for FUNCTION...with
the two developing different feature sets, leading to a greater
mental tax on users to understand these differences.

This implements a first step moving away from the idea that SET-WORD!
with no annotation will provoke locals in functions.  It is not ideal,
as it treats LET as something of a keyword to signal the same kind of
gathering that SET-WORD!s did.  Ideally it would be more of a runtime
concept, so that a LET would actually have to be *running* in order
to create a runtime addition to a "wave" of context:

    data: [let x: 10]
    code: [let y: 10 print y]
    do code  ; this invocation would incarnate y
    ; but since data never runs, x doesn't get allocated

The methodology needed to do that hasn't been fully figured out yet,
but there are several pieces of what would be needed in place.