MSVC 2019 Warnings Update, Rebmake changes
Because the maximal warning level is used, each new compiler release
introduces new warnings that may or may not be interesting.  This
triages those issues for the transition to MSVC 2019.

Also tweaks to rebmake to build files that are taken by MSVC2019
directly.  It mentions the necessary patches if someone felt it
necessary to make it possible to still build for VC2017...which would
be fairly trivial.