Switch from storing underlying to just *checking*
In specific binding, it was observed that there was a negative aspect
to storing a derived function's paramlist in the bound words of a
function's relativized body.  If a function were COPY'd, the new
copy would have its own identity (a new paramlist) but it would want
to reuse the body.  So the concept of storing and checking the
underlying function came about; since the frames were the same size,
it didn't really matter.

With the concept of extending frames on the horizon, this doesn't work,
because the underlying frame doesn't have as many cells as the now
bigger frame.  Since really it is only a debug check to make sure the
specifier is a legal match for the stored binding in the word, we can
twist this around so the stored paramlist is for the function we want,
but then relax the check to make sure the underlying functions match.