Fix LETs in nested blocks
There was a bug that was making LETs not work in nested blocks unless
the LET was the first item in the block.  This adds the necessary
incrementation in the loop to get the correct behavior.

Also makes LETs actually act invisible, and allows you to use LET
with a plain WORD! by quoting it skippably.  (This is *not* the long
term implementation plan for LET via virtual binding...)