Permit VOID! assignments via SET-WORD! / SET-PATH!
Rebol historically would disallow assignments of the "unset!" value,
e.g. the result of operations like PRINT which had no real intended
return value:

    rebol2>> x: print "like this"
    like this
    ** Script Error: x needs a value

While this provides some amount of safety, it is an impediment to
generalized code, such as:

    x: do block

If you do not know what the block would return, you would wind up
having to use SET/ANY:

    set/any 'x do block

This became even worse with SET-WORD! collecting, e.g.

    r3-alpha>> function [...] [set/any (quote x:) do block]

With LET coming on the scene, it would be returning the value of the
assignment:

    all [
        let pos: find data item
        ...
    ]

Plugging these things together seems to become a mess.  And the value
proposition of giving the errors is somewhat questionable, compared
to just letting the errors come downstream when a void variable is
actually used.

In the interest of providing a "freedom from" value assignments that
don't work, this goes ahead and liberates assignment to always work;
e.g. a SET-WORD! acts as SET/ANY and a GET-WORD! acts as GET/ANY.