Add "Lit-Branching" Experiment
This makes @word, @pa/th, and @(gr o up) constructs legal as a literal
form in branch slots.  The word and path forms are essentially just
abbreviations for putting the word to look up in a block, e.g.:

    if condition @x  =>  if condition [:x]

It is lighter because it avoids a series allocation, and can perform
just a raw fetch of a variable without bringing in the evaluator or
making a new frame.

The @(gr o up) form is interesting because it acts much like a plain
GROUP! does, in that its job is to generatively provide the code for
the branch.  However, the difference is that it will only run that
generative process if the branch should be run.  e.g.

    >> if true (print "Hello" [1 + 2])
    Hello
    == 3

    >> if false (print "Hello" [1 + 2])
    Hello
    ; null

    >> if true @(print "Hello" [1 + 2])
    Hello
    == 3

    >> if false @(print "Hello" [1 + 2])
    ; null

Note the difference in the last case.  Not only can this help avoid
unwanted side-effects in code-generating cases, instances that do
not have side-effects can avoid doing wasted work.

There is no meaning given to @[...] at this time.  While a novel use
could be thought of for it, the plans for this semi-redundant form
(non-evaluative BLOCK!) is to have it usually serve as representation
for datatypes.