Eliminate quoted forms of AND, OR short-circuit clauses
Pre-dating generalized quoting (and soft-quoted branching, and
@words, @pa/ths, etc.) was an idea to contract the expression of
short-circuit clauses in conditional AND and OR.  So:

    if x and [y] [...]

Could become:

    if x and 'y [...]

It was just an experimental idea to try and reduce the visual weight of
using such clauses with simple variables.

But with the existence of soft-quoted branches, this seems a rather
counterintuitive interpretation.  You would not expect an argument
like 'y to imply the fetching and evaluation of the variable in y.
If anything you would think it would act like this:

    >> true and 'y
    == y

This reimagines AND, OR, and XOR as branching constructs--and makes use
of the new facility to fetch variables by literal value:

    >> y: 10

    >> true and @y
    == 10

    >> false and @y
    == #[false]