Merge assert-fixes.h into the reb-c.h header
reb-c.h contains a laundry list of baseline fixes and macros for the
use of cross platform C-augmented-with-C++.  It doesn't really make
sense for the assert-fixes.h patch to be broken out into its own file;
as it sort of sticks out unnecessarily.  This folds it in with the
rest of the C abstractions.

Also cuts down a bit of verbosity in defending the use of DID, and
avoiding GET_FLAG() routines in favor of making flags pre-shifted.