Preload libr3.worker.js file to use as a blob
There is a synchronous request for the URL of where to get the
.worker.js file from.  However, due to cross-domain security rules,
web workers may not be loaded remotely--even when CORS is enabled on
the worker file.

But as a workaround, a worker can be loaded by way of fetch()-ing a
CORS-enabled URL, and then turning that URL into a blob.
Unfortunately, loading a URL via a fetch()'d blob is an asynchronous
operation...which is a mismatch for the interface used by emscripten
which wants the URL synchronously.

Previously a small patch was applied to the libr3.js file, to tweak
it to use an asynchronous load.  This patch is brittle; and has broken
for the second time in a couple of month stretch, when emscripten
changes their source.

This commit switches to a different approach which prefetches the blob
for the worker before initiating the libr3.js code which will demand
the URL.  It then has the blob already on hand to give back as the
answer to the URL request.