Split out core action derivations, flatten META info
(Note: This commit breaks HELP in AUGMENT for the augmented parameters.
That will be addressed in a future commit.)

This establishes lower-level versions of the function derivations that
do not create any "meta" information (e.g. parameter descriptions
used by HELP) on the new functions.  That makes these routines faster
and more focused.  They are named with a `*` (e.g. `ADAPT*`)

That allows the non-`*` versions that do concern themselves with the
maintenance of the "meta" information to handle it separately.  For the
moment that is done via usermode routines that are written out in a
specific order early in the bootstrap.  (e.g. ADAPT)

The higher level routines are based on ENCLOSE'd variations of the
lower-level versions, hence the first derived function to be wrapped
is ENCLOSE* (so the further wrappers will implicitly pick up their
help information).