Use low-level function derivation for COLLECT's KEEP
One advantage of the new low-level function generators is that they are
faster and cheaper.  So cases where being "light" is more important
than having help should use them.

As a test of doing this, the KEEP function generated by COLLECT is
changed to be built by the fast composition tools.