Patches for ODBC Extension since 2017 Experiments
The last time the ODBC extension was known to be in working order was
in late 2017.  While the code has been kept compiling, there was no
continuous integration of the tests...and it was not considered a
priority due to a belief that no one was using it.

It turns out someone was using it...and also, it may be of interest
to a wider audience given the increasing sophistication of Ren-C.
Plus, since it is now a libRebol client it provides a good real-world
test of the API.

This brings it back up to date to run the %odbc-test.reb script about
as well as it did (likely better, as a configuration of unixodbc was
used hence it is run on Linux with Address Sanitizer and bringing in
some fixes there).  Most of the patches related to the change where
APIs do not suppress evaluation by default, hence rebQ() constructs
have to be incorporated.

Also adds support for regular SQLCHAR instead of SQLWCHAR (though
exactly what that means is hazy...it is interpreted here as UTF-8).