Fix nullptr as meaning Rebol NULL state in va_list APIs
Ren-C had an initial concepts surrounding how to unify the variadic
API calls from C and the evaluation of blocks.  This concept ultimately
became formalized as a "feed".

One aspect of the feed that was that due to the evaulator's need for
lookahead and lookback, it would always be holding a "current" value.
There was no intrinsic need that this current value be resident in an
array if it were being evaluated--it could be a virtual item off to
the side that pretended to splice itself in.  This came in handy for
some forms of application.

However, if you didn't want one of these virtual elements and just
wanted to use the array, the virtual value pointer (called `opt_first`)
could be "opted out of".  Opting out was done by passing C's nullptr.

This became contentious with libRebol's decision to use nullptr as
the user-facing form of the transient null state.  It would be
legitimate to start a variadic list with a nullptr and mean it (well,
not to evaluate it, but if a rebValueQ(nullptr) were used that would
be legitimate since it's quoted).

This rewires the code to make the opt_first part of the block oriented
protocols only, and then the variadic APIs assume that nullptr means
a Rebol NULL.

Includes some reformatting and notes on Detect_Rebol_Pointer.