No "rebEND" usage in ODBC, leverage more libRebol API
As libRebol has become more mature, it becomes more and more possible
to implement everything the ODBC extension needs with it--and not need
to resort to calling the low-level %sys-core.h API.  As a further
step toward this, APIs for picking apart date cells are changed to
instead use Rebol calls, and errors are built as REBVAL* instead of
using the low-level REBCTX* pointers (that aren't part of the API).

Additionally, while being able to build as pedantic C89 is a goal of
the project core...extensions like ODBC need not be held to that.
So this goes ahead and leverages the C99 variadic macros so that
`rebEND` does not need to be put at the end of ever API call.