Fix bad date normalization for MAKE DATE!
When a date was being constructed without a time component, it now uses
a "wacky" number for the time to signal that.  Normalize_Date() did
not know about this, so it was rolling over the large number into the
days and months and years to get completely corrupt dates if no
time component were specified in MAKE DATE!.

Also adds some comments regarding the loss of fractional time.