Workaround numeric range issues with BigNum
In theory it is possible when doing updates to use a C datatype with
a very large range (like a BigNum) even when asking to write to a very
small field (like a byte).  So long as the actual value being written
is in the range (e.g. a BigNum that's only 0-255) the right translation
should happen.  So the extension initially just asked for all integer
writes to use BigNums because that was easiest.  And it seemed to work
on 64-bit Windows.

In practice, there are a few problems.  Some ODBC drivers don't support
BigNum at all (e.g. Oracle).  Some support them, but don't do the
translation (e.g. unixodbc with MySQL); so it works if you write a
C 64-bit integer into a BigNum column, but you get corruption if you
try to write it into any smaller column.

This commit uses a workaround to only request BigNums on an as-needed
basis, and uses a signed 32-bit integer otherwise.  That seems to make
unixodbc with MySQL work all right.

In addition, it also switches around to decipher non-WCHAR fields as
UTF-8.  There may need to be an option to control this, but if it does
work then it suggests that possibly the non-Unicode drivers could be
used to work with UTF-8 data.