Changes to support SQLite with unixodbc
SQLite was giving back a report that columns in the database that
were BINARY or BLOB, or DECIMAL, or native (wide) characters were
simply VARCHAR.  This answer came back from SQLDescribeColW().
Without an accurate answer to "what type is this column" then it is
not possible to automatically give back a corresponding Rebol
datatype.

This appears to be a bug or missing feature in the unixodbc SQLite
driver.  Various workarounds were tried until it was discovered (by
looking at the source) that the ODBC driver did in fact know the
string names of the column types, it just wasn't mapping them back
into ODBC types.  It seems to just be leaking the fact that as far
as it is concerned, the backing buffer for all the types in question
are backed by an arbitrary-sized malloc(), as opposed to a fundamental
C datatype like `float` or `int`.

There is a way to ask for the string names, using SQLColAttribute().
So this uses the workaround that if a column is ever reported as a
SQL_VARCHAR, it is checked via string name and mapped into the right
datatype.  This will slow down SQL_VARCHAR queries a little on non
SQLite databases, but has the advantage of being general and not
affecting columns of other types.

Some other ideas were tried along the way to finding out that this was
possible--and this commit includes some work based on that, which
turned out to be useful for other reasons.  For instance: the model
of using "column binding" turned out to not be able to handle the
case when a column turned out to have a value larger than the buffer
which was given.  There was no way to allocate a bigger buffer and
call again.  This adapts to use the alternative approach of manually
calling SQLGetData(), which can make successive queries after finding
out its initial buffer was too small.

Miscellanous cleanups are included related to a better understanding
of the SQL_C_XXX types.