Improve reporting of "unknown ODBC errors"
The message given in SQLGetDiagRec was vague when the process of giving
back an error itself failed.  There aren't very many options for how
that can fail, but it's helpful to know which of the few possibilities
it actually was.

Also reacts correctly to message truncation (though it should probably
increase the buffer size and call the function again).