Try to work around SQLite WCHAR result from Ascii API
One premise of ODBC is that the APIs come in ASCII and Unicode versions.
If it ends in "W", it uses wide (16-bit, generally) characters.
If the API has no "W" at the end, then it is ASCII.

However, when calling the SQLDescribeCol() function on Windows, the
result for the column's type name was coming back as wide characters.
So a string comparison based on ASCII was not working.  The type
"varchar" simply looked like "v".

This attempts to remedy the issue by using the wide character API.