Add POINTFREE and LAMBDA experiment using it
This is an experimental implementation of POINTFREE in usermode.  It
would be rewritten as a native once its functionality is settled.

The concept of "Point Free" or "Tacit Programming" is that it's a way
of relating functions to each other without mentioning their arguments.

So instead of writing something like:

 append-to-data: function [value] [append data value]

You'd have some way of avoiding the redundancy of having to say "value"
twice. So imagine declaring an identical function with a syntax like:

   append-to-data: pointfree [append data]

That would notice you didn't provide all the parameters to append, so
an implicit parameter to the new function would be added. But what if
you wanted instead something like:

   append-10: function [series] [append series 10]

So maybe you would have some syntax in POINTFREE for that where you put
a placeholder in spots that you wanted implicitly picked up. Maybe
BLANK! as a default:

    append-10: pointfree [append _ 10]