AUGMENT fix for adjusting the frame phase
An attempted optimization would make an augmented function use the same
dispatcher as the function it augmented.  This skipped an important
step of switching the frame phase to one in which the augmented args
were not visible, and prevented (among other things) the ability to
augment a PORT! action...which uses a re-dispatch mechanism that should
not see any unused parameters.

Also fixes some whitespace, comments out broken tests.