Redbol CALL/WAIT, base READ, WRITE, SWITCH on AUGMENT
AUGMENT is designed to enable the expansion of function parameters
in derivations.  One of the textbook cases where this is desirable is
the addition of /DEFAULT to the interface for SWITCH in the Redbol
emulation.  This puts it there, but also bases derivations of READ
and WRITE on it...giving behaviors to the /BINARY refinement.

Additionally, it rigs up a /WAIT refinement on CALL.