Switch to NULL for refinements, non-error on null access
The initial motivation for NULL was to be something that could not be
put into a BLOCK! (hence COMPOSE-ing it in must always mean vaporize,
not put a placeholder in).  Because UNSET! had been toyed with as
a sort of vaporizing value in some contexts, NULL was considered the
"real unset" in this vein.

Evolution showed that the error-causing nature of a variable was best
left to a unique value (VOID!).  Thus NULL became better thought of as
the "real NONE!" in terms of failed operations, and was used for its
falseyness in C, and gradually had its error cases narrowed down.

This commit embraces the idea that NULL variables do not cause errors
when they are accessed.  These "unset" cases are not error-triggering
like the set-to-VOID! cases, which are by contrast called "undefined".