Exploit NULL unused refinements with REF() modification
R3-Alpha used NONE! for unused refinements, and had no idea of a Rebol
NULL state.  Using C's nullptr in an API sense for NONE! values would
thus be broken (imagine for intance wanting to query for whether a
NONE! in a block had a newline flag, but with all NONE!s having the
same identity via nullptr this would not be possible).

This meant the C code which wanted to easily read the usage or non
use of a refinement looked for whether the cell was a NONE! or not,
and produced 1 if it was (TRUE) or 0 if it was not (FALSE).  This
was a macro called `D_REF()` that took the argument by number.  Another
macro called `D_ARG()` would give the actual pointer to the cell on
the data stack.

Ren-C involves many changes and streamlinings.  Firstly to take the
arguments by name using token-pasting macros that are auto-generated
in the prep step.  It reduced the names to just ARG() and REF().  And
the cells in question don't live in volatile memory on the data stack,
but in non-contiguous series memory backing frames that stay alive
during the entirety of a function call.

This adds another streamlining point where REF() uses the fact that
arguments which are Rebol NULL can be "nullified" to C's nullptr with
no loss of information.  That nullptr in turn is then also falsey.
Hence it can be tested in conditional expressions or with `not`, but
also be a const pointer to the cell.  This is helpful in natives that
want to deal with refinements (or optional arguments in general) that
are passed to the libRebol API...because that API does not accept
"nulled cells" (which may have different transient addresses), only
the 0 nullptr value of C.

A consequence of this change is that for boolean type checking, to
avoid warnings REF(...) must be passed through a `did` or a `not` to
get explicit conversion to `bool` for C++ or C99 builds, in those
cases where it's being passed as a boolean argument.  This is not
strictly necessary, but it appeases the compiler and leaves the warning
intact.