Allow NULL in PATH! for calling ACTION!
Allows things like:

    append/(if false [/only]) [a b c] [d e]

In accordance to "void is undefined, null is unset" this goes with an
increased tolerance of the null state.  Previous fears like that
`some-func/(:some-mispeling)` would silently be accepted are mitigated,
as that would be a VOID! and not a NULL...and a GET-WORD! wouldn't be
needed for a NULL case anyway.