Bootstrap change for refinements / DEFINED?
Because variable access has fundamentally changed to permit plain
WORD! access for NULL, older executables cannot be hooked to support
NULL refinements.  They must continue to use BLANK!.

This re-engineers the OPT used in the bootstrap process so that if
you give it a NULL it will not produce a void.  That way a BLANK!
refinement can be turned into a NULL in the old build, while a NULL
refinement will be left NULL in the new build.

Also adds DEFINED? and UNDEFINED? in older builds as synonyms for
SET? and UNSET?.