Resolve behavior of enfixed 0-arity functions
Enfixed 0-arity functions are reimagined as functions that bind their
evaluation to be in the same single step as what is on their left.
The exception is if they `<skip>` their left hand side, in which case
they defer and run in the next evaluation step as if they were a
prefix function--setting the argument to null.  Full discussion:

https://github.com/metaeducation/ren-c/issues/581