Fix UTF-8 Everywhere Bookmark Updating Bug
The strategy for making UTF-8 Everywhere perform reasonably for random
seeking in strings is to cache "bookmarks" which map from codepoint
index to byte address along with the series, and then assume that most
random accesses have some amount of locality to where this can speed
things up enough to be tolerable.  (Algorithms requiring true random
access should thus convert the strings to a BLOCK! of codepoints and
back.)

This means all modifications to strings have to either update the
bookmarks if necessary, or throw away the bookmark and have it be
regenerated the next time a seek happens in the string.  (In order to
get the UTF-8 Everywhere branch merged in, the code is committed with
bookmarks probably thrown away more than they should be.)

Currently only one bookmark is maintained per series, and only if it
is considered to be of "significant size" (at the moment, bigger than
the size of a value cell).  This size varies on 32-bit and 64-bit
platforms.

An updating bug on insertions was considering the bookmark to need
updating if the insertion point was *at* the bookmark, when really it
needed to be updated if the insertion point was anywhere *after* the
bookmark.  This amounted to the use of a >= where a > was intended.