Make NEW-LINE/ALL set flag on the tail of the array
Newline markers live conceptually *between* values, but R3-Alpha would
conceptually put them on cells...and this means that for an N element
array there were only N newline markers, when N+1 were needed.  Ren-C
added the feature of tracking that extra bit on the array itself, as
the flag for the array tail.

When setting this bit was added to the NEW-LINE routine, it was done
in a way that didn't add account for the tail when /ALL was used.
This changes it to account for that.