Add Windows Origin Tracking Ability
On Linux, both Valgrind and Address Sanitizer provide mechanisms that
have been used by Panic_Series() to dump out stack traces of where
a series was allocated (as well as the stack at the location of the
panic).  The Windows tool "Dr. Memory" does not provide this feature.
That makes tracking down bugs on Windows more difficult.

This adds the ability to use the `dbghelp.h` API to build stack
traces and save pointers to them in series nodes.  It is slower than
the comparable feature in Address Sanitizer, but can be instrumental
in tracking down Windows-only bugs.

The code for this debug-only feature is in C++ for expedience.