Fix leaked wide string in Windows Error generation
When GetLastError() is used on Windows to get an error code, there is
a provisional libRebol API to turn the resulting DWORD into an ERROR!
value.  This API uses Windows features to get a UTF-16 string out of
the FormatMessage() routine, which it turns into a TEXT! that is
a property of the error.

The API handle for this generated text was not being released.  That
fact had gone unnoticed, because most of the time this form of error
generation was followed by FAIL-ing on that error.  When failing on
such an error, all non-released handles up through the frame of the
TRAP are automatically GC'd.

A rare case where the error was being generated and then poked into
a PORT! field (but not raised via FAIL--at least not immediately)
revealed the oversight.

This error was discovered using the new windows stack tracing ability
from the previous commit.