Fix NEW-LINE markers on paths
Paths were not having their NEW-LINE markers loaded correctly in the
scanner.  This was due to use of a recursive scan, where newline status
was stored in the scan state but overridden in the recursion.

Easiest fix is to just save the newline state before the recursion and
restore it afterward.