Attempt to simplify/clarify Travis build a bit
This makes the bootstrap building process a little more obvious by
naming the bootstrap r3-make "r3-one" and "r3-two", and putting them
in the PATH.  This makes it harder to get mixed up or accidentally
use the wrong build executable if there is a failed step.

Updates some comments, and other light simplifications.