Some tidying up of TLS code
Some of the TLS code had a reasonable amount of "spec-driven" design
(e.g. a state transition table, using the same names as the RFC).
However other parts lacked abstraction and were doing some fairly
manual and ad-hoc extraction of binary data via AT and SKIP, as
opposed to making it clearer how the protocol bytes are decoded.

This takes a small step toward making that more coherent, by using
a new "GRAB" function.  This function copies part of a binary and
uses the SET-WORD! on its left hand side for error messages if the
desired decoding size is not present.

It is a placeholder for a more declarative "binary parse" style
abstraction--but is a step forward in seeing what's going on.