Revise Ren-C's rules for CR and 0 bytes in strings
This commit tries to build on Rebol's historical concept of unifying
strings within the system to use LF-only.  But rather than try "magic"
to filter out CR LF sequences (and "magically" put them back later), it
adds in speedbumps to try and stop CR from casually getting into
strings.  Then it encourages active involvement at the source level
with functions like ENLINE and DELINE when a circumstance can't be
solved by standardizing the data sources themselves:

https://forum.rebol.info/t/1264

It also revises Ren-C's rules to be  prescriptive about disallowing 0
bytes in strings, to more safely use the rebSpell() API, which only
returns a pointer and must interoperate with C.  It enforces the use of
BINARY! if you want to embed 0 bytes (and using the rebBytes() API,
which always returns a size.)