Fix REMOVE in strings
Before UTF-8 Everywhere, the REMOVE was a common series action that
could operate on a unit-level basis for the series.  However, once
variable-sized codepoints were introduced, the action had to be
handled distinctly in a way that might remove several units (bytes)
for a single conceptual item (codepoint).

This change didn't take into account that the default /PART for
REMOVE is 1, and would thus act like a CLEAR on strings.

This fixes that issue and adds a bare minimum test on this point.

A better fix might rethink REMOVE as being built on top of CHANGE with
NULL as the default material to "change to".