Allow soft failure from TAKE
Concern about making sure that users "knew what they were doing" when
TAKE-ing from an empty block (and having no side effects) led to a
different `TAKE*` (empty block tolerant, returning NULL) from TAKE
(error on empty block).

This particular safety mechanism--which adds a "tic" of a `*` for a
common desire--seems of questionable value given current thinking about
how the out-of-the-box constructs should work.  Those who need safety
are encouraged to build their own.

Additionally, having a unique state to represent the inability to take
anything (NULL) that cannot possibly exist in the block makes it seem
less dangerous on the whole...as most TAKEs will use the value that
is removed.