Don't bother with 128-bit math in easy-ecc code
The math file used for elliptic curve cryptography tried to take
advantage of 128-bit integer support if it was present, for its BigNum
implementation.

This was causing various warnings and build setting headaches.  While
they might be possible to work through, the importance of doing so is
very low in the priority scheme.  This is especially true since having
every cryptographic algorithm linked into a Rebol executable
reimplement BigNums is undesirable.

For now this just says that there's no attempt to try to use 128-bit
integers (which are in no C standard).