Use a TEXT! Rebol value for POSIX string buffer
It is desirable to be able to hook the terminal on a per-key basis
(when that ability is available).  If this is done, then it will also
be desirable to get the state of the input as a Rebol value.

This makes another incremental step toward that goal by changing the
POSIX readline C code to store its state in a Rebol TEXT!.