Split POSIX console into "event" and "reaction" parts
This furthers the process of modularizing the posix termios code, to
clearly deliniate the "keys/instructions" from the corresponding
behaviors.

The key-returning part uses Rebol values as their currency (though
notably not key EVENT!s, just basic WORD!s and CHAR!s that capture what
happened).

The reaction part leans more heavily upon libRebol, in order to ease
the eventual transition to where these reactions are customizable in
usermode, as a more deeply-hooked level of terminal.  This terminal
should be able to react to keystrokes without showing them on the
screen, or to be interruptable for doing calculations while awaiting
input.  This is a step toward that goal.